# packages/request/api-spec.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","summary":"用户注册","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"注册成功"},"400":{"description":"注册失败"},"409":{"description":"用户名/手机号/邮箱已存在"}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"登录成功"},"401":{"description":"未授权"}},"tags":["Auth"]}},"/auth/login/sms":{"post":{"operationId":"AuthController_loginWithSms","summary":"短信验证码登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SmsLoginDto"}}}},"responses":{"200":{"description":"登录成功"},"400":{"description":"登录失败"}},"tags":["Auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshTokens","summary":"刷新令牌","parameters":[],"responses":{"200":{"description":"刷新成功"},"401":{"description":"刷新失败"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","summary":"退出登录","parameters":[],"responses":{"200":{"description":"退出成功"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","summary":"获取当前用户信息","parameters":[],"responses":{"200":{"description":"获取成功"},"401":{"description":"未授权"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/calendar/solar-terms":{"get":{"operationId":"SolarTermsController_getSolarTerms","summary":"获取二十四节气","parameters":[{"name":"year","required":false,"in":"query","description":"年份，默认为当前年份","schema":{"example":2024,"type":"number"}}],"responses":{"200":{"description":"返回二十四节气数据","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolarTerm"}}}}}},"tags":["SolarTerms"],"security":[{"bearer":[]}]}}},"info":{"title":"Yuzhi API","description":"The Yuzhi API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"johndoe","description":"用户名"},"email":{"type":"string","example":"john@example.com","description":"电子邮件"},"phoneNumber":{"type":"string","example":"13800138000","description":"手机号码"},"password":{"type":"string","example":"Password123!","description":"密码"}},"required":["username","email","phoneNumber","password"]},"LoginUserDto":{"type":"object","properties":{"usernameOrPhone":{"type":"string","example":"johndoe","description":"用户名或手机号"},"password":{"type":"string","example":"Password123!","description":"密码"},"rememberMe":{"type":"boolean","example":true,"description":"是否记住登录状态"}},"required":["usernameOrPhone","password"]},"SmsLoginDto":{"type":"object","properties":{"phoneNumber":{"type":"string","example":"13800138000","description":"手机号码"},"code":{"type":"string","example":"123456","description":"验证码"}},"required":["phoneNumber","code"]},"SolarTerm":{"type":"object","properties":{"id":{"type":"string","description":"主键ID"},"pub_year":{"type":"number","description":"年份"},"name":{"type":"string","description":"节气名称"},"pub_date":{"type":"string","description":"节气公历日期"},"pri_date":{"type":"string","description":"节气农历日期"},"pub_time":{"type":"string","description":"节气时间"},"des":{"type":"string","description":"节气简介"},"youLai":{"type":"string","description":"节气由来"},"xiSu":{"type":"string","description":"节气习俗"},"heath":{"type":"string","description":"节气养生建议"}},"required":["id","pub_year","name","pub_date","pri_date","pub_time","des","youLai","xiSu","heath"]}}}}