# packages/request/src/generated/models/CreateUserDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateUserDto = {
    /**
     * 用户名
     */
    username: string;
    /**
     * 电子邮件
     */
    email: string;
    /**
     * 手机号码
     */
    phoneNumber: string;
    /**
     * 密码
     */
    password: string;
};