# packages/request/src/generated/models/LoginUserDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LoginUserDto = {
    /**
     * 用户名或手机号
     */
    usernameOrPhone: string;
    /**
     * 密码
     */
    password: string;
    /**
     * 是否记住登录状态
     */
    rememberMe?: boolean;
};