# packages/request/src/generated/models/SolarTerm.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SolarTerm = {
    /**
     * 主键ID
     */
    id: string;
    /**
     * 年份
     */
    pub_year: number;
    /**
     * 节气名称
     */
    name: string;
    /**
     * 节气公历日期
     */
    pub_date: string;
    /**
     * 节气农历日期
     */
    pri_date: string;
    /**
     * 节气时间
     */
    pub_time: string;
    /**
     * 节气简介
     */
    des: string;
    /**
     * 节气由来
     */
    youLai: string;
    /**
     * 节气习俗
     */
    xiSu: string;
    /**
     * 节气养生建议
     */
    heath: string;
};